requirements.txt
sentence-transformers
qdrant-client
langchain-text-splitters
groq
anthropic
python-dotenv
ollama
